mostly changes in the gallery page
