--- src/pages/explore/index.module.scss
@@ -1,5 +1,0 @@
-.grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(16rem, 1fr));
-  grid-gap: 1rem;
-}
