+++ src/pages/gallery/[...slug].tsx
